refactor: extract AgentOnboardingLaunchPanel from ConnectUsingAgent
